Added index.php and changed the Conversion class
The index.php is where I built simple form and instantiated the objects to test if the methods work correctly. I tweaked the Conversion class, so it accepts only the temperature value in the constructor. The toUnit variable that was in the constructor is now passed to the Convert function. This allows to reuse the same object if another conversion is needed. (For example, if we wanted to calculate and display the conversions from celsius to both kelvin and fahrenheit simultaneously)